Remove explicit anchors in Popover, fix hella type issues
