+++ packages/react/src/components/Checklist/FloatingStep.tsx
@@ -1,4 +1,4 @@
-import { useMemo, useRef } from 'react'
+import { useRef } from 'react'
 import * as Popover from '@/components/Popover'
 
 import { Card } from '@/components/Card'
@@ -13,7 +13,6 @@ import { floatingTransitionCSS } from '@/components/Checklist/Floating.styles'
 
 // TODO: Type props
 export function FloatingStep({ onPrimary, onSecondary, openStepId, setOpenStepId, step }) {
-  const anchorId = useMemo(() => `floating-checklist-step-${step.id}`, [step.id])
   const anchorPointerEnterTimeout = useRef<ReturnType<typeof setTimeout>>()
   const { handlePrimary, handleSecondary } = useStepHandlers(step, { onPrimary, onSecondary })
 
@@ -54,15 +53,15 @@ export function FloatingStep({ onPrimary, onSecondary, openStepId, setOpenStepId
   const { videoProps } = getVideoProps(step.props ?? {})
 
   return (
-    <>
-      <Flex.Row
+    <Popover.Root align="start" open={isStepOpen} side="right" sideOffset={8}>
+      <Popover.Trigger
+        as={Flex.Row}
         alignItems="center"
         borderRadius="md"
         gap="2"
         justifyContent="space-between"
         onPointerEnter={handlePointerEnter}
         onPointerLeave={handlePointerLeave}
-        id={anchorId}
         padding="1 2"
         part="floating-checklist-step-list-item"
         style={{
@@ -72,48 +71,40 @@ export function FloatingStep({ onPrimary, onSecondary, openStepId, setOpenStepId
       >
         <Text.Body2 part="floating-checklist-step-title">{step.title}</Text.Body2>
         <CheckIndicator checked={step.$state.completed || step.$state.skipped} size="18px" />
-      </Flex.Row>
-      <Popover.Root
-        align="start"
-        anchor={`#${anchorId}`}
-        open={isStepOpen}
-        side="right"
-        sideOffset={8}
-      >
-        <Popover.Content css={floatingTransitionCSS} part="floating-step">
-          <Card
-            backgroundColor="neutral.background"
-            border="md solid neutral.border"
-            borderRadius="md"
-            gap="3"
-            minWidth="400px"
-            p="2"
-          >
-            <Card.Media
-              src={step.videoUri ?? step.imageUri}
-              type={step.videoUri ? 'video' : 'image'}
-              css={{ objectFit: 'contain', width: '100%' }}
-              {...videoProps}
-            />
-            <Card.Header dismissible={false} padding="0 1" subtitle={step.subtitle} />
+      </Popover.Trigger>
+      <Popover.Content css={floatingTransitionCSS} part="floating-step">
+        <Card
+          backgroundColor="neutral.background"
+          border="md solid neutral.border"
+          borderRadius="md"
+          gap="3"
+          minWidth="400px"
+          p="2"
+        >
+          <Card.Media
+            src={step.videoUri ?? step.imageUri}
+            type={step.videoUri ? 'video' : 'image'}
+            css={{ objectFit: 'contain', width: '100%' }}
+            {...videoProps}
+          />
+          <Card.Header dismissible={false} padding="0 1" subtitle={step.subtitle} />
 
-            <Flex.Row gap={3} justifyContent="flex-end" part="card-footer">
-              <Card.Secondary
-                disabled={step.$state.blocked}
-                onClick={wrappedHandleSecondary}
-                padding="1 2"
-                title={secondaryButtonTitle}
-              />
-              <Card.Primary
-                disabled={step.$state.blocked}
-                onClick={wrappedHandlePrimary}
-                padding="1 2"
-                title={primaryButtonTitle}
-              />
-            </Flex.Row>
-          </Card>
-        </Popover.Content>
-      </Popover.Root>
-    </>
+          <Flex.Row gap={3} justifyContent="flex-end" part="card-footer">
+            <Card.Secondary
+              disabled={step.$state.blocked}
+              onClick={wrappedHandleSecondary}
+              padding="1 2"
+              title={secondaryButtonTitle}
+            />
+            <Card.Primary
+              disabled={step.$state.blocked}
+              onClick={wrappedHandlePrimary}
+              padding="1 2"
+              title={primaryButtonTitle}
+            />
+          </Flex.Row>
+        </Card>
+      </Popover.Content>
+    </Popover.Root>
   )
 }
